Feat create jeu donnee in job web api

+++ JobWebApi/Enities/Logiciels.cs
@@ -27,7 +27,7 @@
     {
         public float Numero { get; set; }
         public string CodeLogiciel { get; set; } = "";
-        public short Millesime { get; set; }
+        public int Millesime { get; set; }
         public DateTime DateOuverture { get; set; }
         public DateTime DateSortiePrevue { get; set; }
         public DateTime? DateSortieReelle { get; set; }
